+++ mitra_config/src/config.rs
@@ -31,6 +31,9 @@ use super::registration::RegistrationConfig;
 use super::retention::RetentionConfig;
 use super::software::SoftwareMetadata;
 
+const DEFAULT_HTTP_HOST: &str = "127.0.0.1";
+const DEFAULT_HTTP_PORT: u32 = 8383;
+
 fn default_log_level() -> LogLevel { LogLevel::Info }
 
 const fn default_web_client_rewrite_index() -> bool { true }
@@ -68,10 +71,10 @@ pub struct Config {
     #[serde(default = "default_media_proxy_enabled")]
     pub media_proxy_enabled: bool,
 
-    pub http_host: Option<String>,
-    pub http_port: Option<u32>,
+    http_host: Option<String>,
+    pub(super) http_port: Option<u32>,
     // Overrides http_host and http_port
-    pub http_socket: Option<String>,
+    http_socket: Option<String>,
     // Unix socket permissions (example: 0o640)
     pub http_socket_perms: Option<u32>,
 
@@ -152,16 +155,19 @@ impl Config {
     }
 
     pub fn http_socket(&self) -> String {
-        match (&self.http_socket, &self.http_host, self.http_port) {
-            (Some(http_socket), _, _) => http_socket.clone(),
-            (None, Some(http_host), Some(http_port)) => {
+        match &self.http_socket {
+            Some(http_socket) => http_socket.clone(),
+            None => {
+                let http_host =
+                    self.http_host.as_deref().unwrap_or(DEFAULT_HTTP_HOST);
+                let http_port =
+                    self.http_port.unwrap_or(DEFAULT_HTTP_PORT);
                 if http_host.parse::<Ipv6Addr>().is_ok() {
                     format!("[{http_host}]:{http_port}")
                 } else {
                     format!("{http_host}:{http_port}")
                 }
             },
-            _ => panic!("either http_socket or http_host and http_port must be specified"),
         }
     }